add regex
