+++ opentelemetry-api/src/opentelemetry/util/__init__.py
@@ -11,6 +11,7 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
+import re
 import time
 from logging import getLogger
 from typing import Sequence, Union
@@ -51,10 +52,17 @@ def _load_provider(
         logger.error("Failed to load configured provider %s", provider)
         raise
 
+# Pattern for matching up until the first '/' after the 'https://' part.
+_URL_PATTERN = r"(https?|ftp)://.*?/"
 
-def disable_tracing_path(url_path: str, excluded_paths: Sequence[str]) -> bool:
-    for path in excluded_paths:
-        if url_path.startswith(path):
+
+def disable_tracing_path(url: str, excluded_paths: Sequence[str]) -> bool:
+    print(url)
+    if excluded_paths:
+        # Match only the part after the first '/' that is not in _URL_PATTERN
+        regex = "{}({})".format(_URL_PATTERN, "|".join(excluded_paths))
+        print(regex)
+        if re.match(regex, url):
             return True
     return False
